Fixing actions errors.
Current Way:
Set specific action variables to individual actions throughout civ/cop/gov.
If need to be remove, remove all action variables.

This is bad, as all have to be used, or sqs will just puke, as well as mislabling will guarentee it will not be removed correctly.

New Way:
Each action added gets placed in a list.
If need to be remove, remove the list.

This means no longer having to worry about using specific variables in all scripts.